LPS-54509 Extract inner test classes from EntityCacheImplTest.
